+++ examples/nuxt-ssr-example/nuxt-quill-plugin.js
@@ -1,7 +1,5 @@
 
 import Vue from 'vue'
+import VueQuillEditor from 'vue-quill-editor/dist/ssr'
 
-if (process.browser) {
-  const VueQuillEditor = require('vue-quill-editor/dist/ssr')
-  Vue.use(VueQuillEditor)
-}
+Vue.use(VueQuillEditor)